Refactor error fixing/analyis code into its own module
